requirements.txt
flask
flask-cors
pandas
scikit-learn
yfinance
requests
python-dotenv
gunicorn